fix: modify migrate link reference to not accidentally replace [] values (#1220)
[![PR App][icn]][demo] | Fix RM-14308
:-------------------:|:----------:

## 🧰 Changes

We've had a migrated project have its document content malformed because
a [] is recognized as a link & accidentally replaced with another
content.

Update the migrate-link-reference logic to not use positions.

## 🧬 QA & Testing

Added some tests to check if we accidentally modify [] blocks. Let me
know if it’s possible for link blocks to not be [].

This solves the issue for the markdown in
https://linear.app/readme-io/issue/RM-14308/page-content-not-converting-correctly

- [Broken on production][prod].
- [Working in this PR app][demo].


[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg