fix: ignore legacy magic blocks when stripping comments (#1228)
[![PR App][icn]][demo] | RM-14371 |
:------------------|---:

> [!IMPORTANT] 
> I have a bandaid provisional fix for this issue up here
https://github.com/readmeio/readme/pull/16356 to relieve the pressure on
getting this figured out.

This PR attempts to address the `@readme/markdown` portion of
https://linear.app/readme-io/issue/CX-2524/legacy-customer-sumsub-has-pages-with-large-html-blocks-not-loading
where a few legacy projects with very large html blocks aren't loading.

This is a regression from https://github.com/readmeio/readme/pull/16342
In the readme app, we strip html comments from a document's markdown
body. We do this on the server _before_ the markdown is used in the
`rdmd` render. [See this
diff](https://github.com/readmeio/readme/pull/16342/files#diff-c272ef1c0b573e88f6fc0d9548236cd5d114cab6731f4499cbad12a7ddb7bf14)

The function we use to strip html comments is a plain `remarkParse`
pipeline that handles either markdown or mdx content:
https://github.com/readmeio/markdown/blob/39cb0129ff78c0a9c53f3ff332bb057b506c4db2/lib/stripComments.ts#L15-L24

I think there's something going on with the pipeline modifying the magic
block syntax, but we've only seen it on html blocks with tons of code
within them, so it's hard to pinpoint. I think the best solution is to
just skip magic blocks altogether in the parsing.

I made an attempt at doing this in the legacy package but couldn't find
a clear path. [Here's the PR with my attempt at
that.](https://github.com/readmeio/markdown/pull/1226/files) We don't
want to actually transform the magic blocks into HTML -- we want to
alter the surrounding markdown and return it with the magic blocks
untouched. Not sure how to do that with an ast without a lot of
complexity! But maybe I'm missing something obvious.

So the best I could do is update the `stripComments` function in
`@readme/markdown` to skip magic block syntax. It was much easier to
just use regex to extract the blocks before the remark pipeline runs and
then restore them afterwards.

## 🧰 Changes
- Modify `stripComments` to use a new `extractMagicBlocks` util that
swaps magic blocks with placeholders before parsing the markdown. After
the markdown is parsed the placeholders are restored with the original
magic block text.

## 🧬 QA & Testing
Do the tests pass? I did link a markdown build from this branch to my
readme repo locally and it fixed the issue on clones of the projects in
question.

- [Broken on production][prod].
- [Working in this PR app][demo].


[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg

---------

Co-authored-by: Daniels Lee <daniels@readme.io>