fix(stripComments): ensure extraction token is parsed block level (#1246)
[![PR App][icn]][demo] | 
:-------------------:|

Fixes an issue we found with the `stripComments` function where magic
blocks immediately following a list would recieve an indentation,
causing our rdmd parser not to render them as the correct element.

The `stripComments` function uses a remark pipeline to transform text
into mdast, then remove any "comment" node types, then serializes the
mdast it back to plain md. In the readme app we use this to sanitze
comments from documents, and the sanitized md then gets rendered with
rdmd.


https://github.com/readmeio/markdown/blob/0b855933dc24726142d822ba5e56b1507e6dc190/lib/stripComments.ts#L15-L48

The `stripComments` pipeline has a pre-processing step that extracts all
magic block syntax and replaces it with a special token in the md. Then
after the pipeline runs we restore the tokens with the magic block
syntax.

The token has special characters to make it unique. To ensure remark
doesn't attempt to parse those special chars as markdown, we wrap the
token in backticks so remark just sees it as a code element.


https://github.com/readmeio/markdown/blob/0b855933dc24726142d822ba5e56b1507e6dc190/lib/utils/extractMagicBlocks.ts#L24-L26

But the code element is an inline element -- when this token appears
immediately after a list element without a newline, remark indents the
token as if it belongs to the content of the previous list item.

| Before | After |
| --- | --- |
| <img width="307" height="232" alt="image"
src="https://github.com/user-attachments/assets/8a44c828-a0a0-4d95-a46f-99235c9ec92a"
/> | <img width="433" height="185" alt="image"
src="https://github.com/user-attachments/assets/8b6d30fe-d772-4a9e-ad6f-fe677ea7d10f"
/> |


## 🧰 Changes
- Add a `\n` newline to the magic block extraction token so remark
always treats it as a block level element. Then when it gets restored
with its magic block value, rdmd will render it properly.


## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].


[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg