fix: downgrade `mdast-util-mdx-expression` (#1251)
[![PR App][icn]][demo] | Fix RM-XYZ
:-------------------:|:----------:

## 🧰 Changes

We were seeing a weird minor regression when publishing the new `mdxish`
engine. The new engine used `mdast-util-mdx-expression` v2.0.1 instead
of v2.0.0 which was used as a dependency for the old `mdx` engine (via
`remarkMdx` if not mistaken). While the `mdx` pipeline remained
untouched, using v2.0.1 in `mdxish` unintentionally caused the old `mdx`
engine to also use this newer version... this introduced a slight _bug
(Im not sure if this was intentional)_ where it would indent
expressions, including JSX comments when de- or serializing.

What would usually be something like
```
{/*

## Hey-o

*/}
```

would become 

```
{/*

  ## Hey-o

  */}
```

This is a short demo showcase of the bug:


https://github.com/user-attachments/assets/fe11a82a-01c4-4478-b147-d684a8e5dacb

Ive managed to trace the origin of this behaviour to [this
commit](https://github.com/syntax-tree/mdast-util-mdx-expression/commit/8e5d29000db2fc18e73bb8baa76dd4da8a72b2ce)
WITHIN the plugin library itself. Reverting back to v2.0.0 seems to work
for now. Ive added a check against this to make sure this does not
happen in the future.

> [!IMPORTANT]
> We need to make sure future work regarding `mdx` or `mdxish` uses the
correct v2.0.0 version of `mdast-util-mdx-expression`

## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].


[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg