fix(sanitizeComments): handle magic block on the first line of document (#1276)
[![PR App][icn]][demo] | 
:-------------------:|

Engine-side fix for [this
ticket](https://linear.app/readme-io/issue/CX-2632/akamai-seeing-magic-block-0-where-an-embedded-video-should-be)

Customer is seeing `__MAGIC_BLOCK_0__` where an embedded video should
be.

The issue is they're adding it as a `<video>` element within a Custom
HTML block on the very first line of the document.

You can reproduce this by adding any magic block syntax to the very top
of a document on a legacy project with the sanitize comments flag:
```
[block:html]
{
  "html": "<div>Ooooops!</div>"
}
[/block]
```


https://github.com/user-attachments/assets/076ed306-b78d-429d-b5d1-728d059d0acf


## 🧰 Changes
- Add a special check for a magic block token that has had it's
preceding `\n` trimmed at the very beginning of a document.

## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].


[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg