feat(mdxish): support custom component tag extraction with mdxish syntax (#1277)
[![PR App][icn]][demo] | 
:-------------------:|

Engine-side fix for
https://linear.app/readme-io/issue/CX-2651/custom-mdx-components-are-broken

Gitto calls the `tags` lib function to get all of the custom blocks on a
page:

https://github.com/readmeio/markdown/blob/8cf4e217bdcbd3f152fa5484d49ab6f0b903b868/lib/tags.ts

But this uses a `remark-mdx` processor, which doesn't handle mdxish
syntax, specifically magic blocks. We need a new `tags` function
specifically for handling mdxish syntax.

## 🧰 Changes
- Adds `mdxishTags` lib function
- Add unit tests to cover existing `tags` function extraction behavior
as well as correct handling of magic block syntax

## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].


[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg